Split files based on functions. Added Task and Risk functions;
