+++ server/app/test.html
@@ -39,14 +39,18 @@ Request:
 <option value="projectread">		Read Project</option>
 <option value="projectupdate">		Update Project</option>
 <option value="projectdelete">		Delete Project</option>
-<option value="task">Task</option>
-<option value="risk">Risk</option>
+<option value="tasknew">			New Task</option>
+<option value="risknew">			New Risk</option>
 </select><br>
 Username: <input type="text" name="usern"><br>
 Project Name: <input type="text" name="projn"><br>
+Task Name: <input type="text" name="taskn"><br>
+Risk Name: <input type="text" name="riskn"><br>
 Field: <input type="text" name="field"><br>
 Value: <input type="text" name="val"><br>
 Project ID: <input type="text" name="projId"><br>
+Task ID: <input type="text" name="taskId"><br>
+Risk ID: <input type="text" name="riskId"><br>
 <input type="button" onclick="runQuery()" value="Run" name="Button">
 </form>
 <h3>Request</h3>
